% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_S3.R
\name{find_S3class}
\alias{find_S3class}
\title{Find S3 classes and methods}
\usage{
find_S3class(S3_name)
}
\arguments{
\item{S3_name}{The name of the class to be created}
}
\description{
These functions help finding the class and function definitions for S3 classes
in the active project.
}
