% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_S3generic.R
\name{new_S3generic}
\alias{new_S3generic}
\title{Create a new S3 generic}
\usage{
new_S3generic(name, ..., S3_output = "clipboard", S3_overwrite = FALSE)
}
\arguments{
\item{name}{The name of the new generic}

\item{...}{The input parameters of the new generic.}

\item{S3_output}{Control the output of the function: Either 'clipboard' or
'file' or 'console'. Default 'clipboard'}

\item{S3_overwrite}{Control whether to overwrite an existing file. Default
FALSE.}
}
\description{
These functions help set up the base structure for a new S3 generic
}
