% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_S3method.R
\name{new_S3method}
\alias{new_S3method}
\title{Define new S3 methods}
\usage{
new_S3method(S3_generic, S3_name, S3_output = "clipboard")
}
\arguments{
\item{S3_generic}{A generic function. Enter as a call to that function,}

\item{S3_name}{The name of the class for which this method is written
without parantheses: S3_generic = myfun}

\item{S3_output}{Control the output of the function: Either 'clipboard' or
'file' or 'console'. Default 'clipboard'}
}
\description{
This function helps you to generate new methods for existing generics. It
automatically applies a structure that matches the generic.
}
